+++ src/picker/components/Picker/framework.js
@@ -1,69 +1,7 @@
-// via https://github.com/component/escape-html/blob/b42947eefa79efff01b3fe988c4c7e7b051ec8d8/index.js
-function escapeHtml (string) {
-  const str = '' + string
-  const match = /["'&<>]/.exec(str)
+import { escapeHtml, getFromMap, parseTemplate, toString } from './utils.js'
 
-  if (!match) {
-    return str
-  }
-
-  let escape
-  let html = ''
-  let index = 0
-  let lastIndex = 0
-
-  for (index = match.index; index < str.length; index++) {
-    switch (str.charCodeAt(index)) {
-      case 34: // "
-        escape = '&quot;'
-        break
-      case 38: // &
-        escape = '&amp;'
-        break
-      case 39: // '
-        escape = '&#39;'
-        break
-      case 60: // <
-        escape = '&lt;'
-        break
-      case 62: // >
-        escape = '&gt;'
-        break
-      default:
-        continue
-    }
-
-    if (lastIndex !== index) {
-      html += str.substring(lastIndex, index)
-    }
-
-    lastIndex = index + 1
-    html += escape
-  }
-
-  return lastIndex !== index ? html + str.substring(lastIndex, index) : html
-}
-
-function toString (value) {
-  if (typeof value === 'undefined') {
-    return 'undefined'
-  } else if (value === null) {
-    return 'null'
-  }
-  return value.toString()
-}
-
-function parseTemplate (htmlString) {
-  const template = document.createElement('template')
-  template.innerHTML = htmlString
-
-  if (process.env.NODE_ENV !== 'production') {
-    if (template.content.children.length !== 1) {
-      throw new Error('only 1 child allowed for now')
-    }
-  }
-  return template
-}
+const domInstancesCache = new WeakMap()
+const unkeyedSymbol = Symbol('un-keyed')
 
 function patchChildren (newChildren, binding) {
   const { targetNode } = binding
@@ -333,6 +271,33 @@ function parseHtml (tokens) {
   }
 }
 
-export {
-  parseHtml
+export function createFramework (state) {
+  let domInstances = getFromMap(domInstancesCache, state, () => new Map())
+  let iteratorKey = unkeyedSymbol
+
+  function html (tokens, ...expressions) {
+    const domInstancesForKey = getFromMap(domInstances, iteratorKey, () => new WeakMap())
+    const domInstance = getFromMap(domInstancesForKey, tokens, () => parseHtml(tokens))
+
+    const { update } = domInstance
+    const { dom } = update(expressions)
+    return { dom }
+  }
+
+  function map (array, callback, keyFunction, mapKey) {
+    const originalCacheKey = iteratorKey
+    const originalDomInstances = domInstances
+    domInstances = getFromMap(domInstances, mapKey, () => new Map())
+    try {
+      return array.map((item, index) => {
+        iteratorKey = keyFunction(item)
+        return callback(item, index)
+      })
+    } finally {
+      iteratorKey = originalCacheKey
+      domInstances = originalDomInstances
+    }
+  }
+
+  return { map, html }
 }
